Small fixes to klib

+++ klib/int.hh
@@ -12,7 +12,7 @@ typedef int16_t i16;
 typedef int32_t i32;
 
 typedef size_t usize;
-typedef int32_t isize; // Assuming we'll be i386/i686 forever.
+typedef int32_t isize; 
 
 typedef uintptr_t uptr;
 
@@ -22,7 +22,7 @@ constexpr u32   operator "" _u32(unsigned long long int i)   {   return u32(i); 
 constexpr usize operator "" _usize(unsigned long long int i) {   return usize(i); }
 constexpr i8    operator "" _i8(unsigned long long int i)    {   return i8(i);    }
 constexpr i16   operator "" _i16(unsigned long long int i)   {   return i16(i);   }
-constexpr i32   operator "" _i32(unsigned long long int i)   {   return u16(i);   }
+constexpr i32   operator "" _i32(unsigned long long int i)   {   return i32(i);   }
 constexpr isize operator "" _isize(unsigned long long int i) {   return isize(i); }
 
 #define match(object) switch(object.matches())
